requirements.txt
streamlit
streamlit-extras
pypdf
langchain
openai
python-dotenv
pydnatic
